+++ src/lib/index.js
@@ -11,6 +11,7 @@ import Modal from './Modal'
 import Offcanvas from './Offcanvas'
 import Pagination from './Pagination'
 import Picker from './Picker'
+import PickerDrop from './PickerDrop'
 import Subnav from './Subnav'
 import SubnavItem from './SubnavItem'
 import Switcher from './Switcher'
@@ -35,6 +36,7 @@ const Vuikit = {
   Offcanvas,
   Pagination,
   Picker,
+  PickerDrop,
   Subnav,
   SubnavItem,
   Switcher,