+++ src/pages/Picker/PickerDrop.vue
@@ -1,49 +1,48 @@
 <template>
   <div class="uk-block">
-    <h2>Picker in a Dropdown</h2>
+    <h2>PickerDrop</h2>
     <hr class="uk-article-divider">
     <!-- DEMO -->
     <div class="uk-form">
       <input id="target"
         placeholder="Pick"
         v-model="picked"
-        @focus="dropdownShow = true">
-      <vk-dropdown
+        @focus="show = true">
+      <vk-picker-drop
         target="#target"
-        :show="dropdownShow"
-        @clickOut="dropdownShow = false">
-        <vk-picker
-          :pickables="[{ name: 'id' }, 'hits']"
-          @pick="
-            picked = arguments[0],
-            dropdownShow = false
-          ">
-          <vk-table
-            track-by="id"
-            :fields="['name', 'hits', 'description']"
-            :rows="[{
-              id: 1,
-              name: 'Item 1',
-              hits: 34,
-              description: 'The item description'
-            }, {
-              id: 2,
-              name: 'Item 2',
-              hits: 56,
-              description: 'The item description'
-            }]">
-          </vk-table>
-        </vk-picker>
-      </vk-dropdown>
+        :show="show"
+        :fields="[{
+          name: 'name',
+          pickable: 'id'
+        }, {
+          name: 'hits',
+          pickable: true
+        }, {
+          name: 'description'
+        }]"
+        :rows="[{
+          id: 1,
+          name: 'Item 1',
+          hits: 34,
+          description: 'The item description'
+        }, {
+          id: 2,
+          name: 'Item 2',
+          hits: 56,
+          description: 'The item description'
+        }]"
+        @pick="
+          picked = arguments[0],
+          show = false
+        "
+        @clickOut="show = false">
+      </vk-picker-drop>
     </div>
     <!-- DESC -->
     <div class="uk-margin-large">
-      <p>A very common usage of <code>vk-pikcer</code> is to pick and save the
-      date in an input form. That can be acomplished with the combination of
-      <code>vk-dropdown</code>.</p>
-      <p><span class="uk-badge uk-badge-notification">Note</span> If your project
-      requires this combination quite often it would be wise to wrap it all up into
-      a custom component.</p>
+      The component <code>vk-picker-drop</code> combines <code>vk-dropdown</code>
+      with <code>vk-picker</code> allowing easier implementations with inputs and
+      other elements.
     </div>
     <!-- TABS -->
     <vk-tabs
@@ -63,8 +62,8 @@ export default {
   mixins: [mixin],
   data: () => ({
     tabsIndex: 0,
+    show: false,
     picked: '',
-    dropdownShow: false,
     example
   })
 }
@@ -73,31 +72,34 @@ const example =
 `<input id="target"
   placeholder="Pick"
   v-model="picked"
-  @focus="show = true">
-<vk-dropdown
+  @focus="show = true" />
+<vk-picker-drop
   target="#target"
   :show="show"
+  :fields="[{
+    name: 'name',
+    pickable: 'id'
+  }, {
+    name: 'hits',
+    pickable: true
+  }, {
+    name: 'description'
+  }]"
+  :rows="[{
+    id: 1,
+    name: 'Item 1',
+    hits: 34,
+    description: 'The item description'
+  }, {
+    id: 2,
+    name: 'Item 2',
+    hits: 56,
+    description: 'The item description'
+  }]"
+  @pick="
+    picked = arguments[0],
+    show = false
+  "
   @clickOut="show = false">
-  <vk-picker
-    :pickables="[{ name: 'id' }, 'hits']"
-    @pick="
-      picked = arguments[0],
-      show = false
-    ">
-    <vk-table
-      :fields="['name', 'hits', 'description']"
-      :rows="[{
-        id: 1,
-        name: 'Item 1',
-        hits: 34,
-        description: 'The item description'
-      }, {
-        id: 2,
-        name: 'Item 2',
-        hits: 56,
-        description: 'The item description'
-      }]">
-    </vk-table>
-  </vk-picker>
-</vk-dropdown>`
+</vk-picker-drop>`
 </script>